Create Circles and Graphs - JavaFX
No one really knows why this is useful but this project uses the JavaFX GUI to display a series of circles on the pane, which the user can add by left clicking.  The colors of the circles match the graph edge which the circles lie on.  As circles begin to touch each other, they join into a common edge and assume the same color as their contacting neighbors.

Click on the pane to add a circle.  Add a neighboring circle in a similar fashion and watch the magic happen......